+++ Sources/KeyboardKit/Previews/KeyboardLayoutProvider+Preview.swift
@@ -24,4 +24,8 @@ public class PreviewKeyboardLayoutProvider: KeyboardLayoutProvider {
     public func keyboardLayout(for context: KeyboardContext) -> KeyboardLayout {
         provider.keyboardLayout(for: context)
     }
+    
+    public func register(inputSetProvider: KeyboardInputSetProvider) {
+        provider.register(inputSetProvider: inputSetProvider)
+    }
 }