+++ RELEASE_NOTES.md
@@ -40,6 +40,7 @@ Besides the new things listed below, there are a bunch of new extensions, images
 * `KeyboardInputViewController` has a new `keyboardSecondaryInputActionProvider`.
 * `KeyboardInputViewController` has a new `keyboardSecondaryInputCalloutContext`.
 * `KeyboardInputViewController` `setup(with:)` is now open and overridable.
+* `KeyboardLayoutProvider` has a new `register(inputSetProvider:)` to simplify changing global input set provider.
 * `Layout` has new types and providers for working with keyboard layouts.
 * `Locale` is a new namespace with a few new locale-specific utils.
 * `LocaleDictionary` is a simple dictionary wrapper to work with localized resources.